Make mock version for testing
