+++ wmexchanger/src/main/scala/org/clulab/wm/wmexchanger/wmconsumer/KafkaConsumerish.scala
@@ -1,0 +1,6 @@
+package org.clulab.wm.wmexchanger.wmconsumer
+
+trait KafkaConsumerish {
+  def poll(duration: Int): Unit
+  def close(): Unit
+}
