+++ wmexchanger/src/main/scala/org/clulab/wm/wmexchanger/wmconsumer/RestConsumerish.scala
@@ -1,0 +1,9 @@
+package org.clulab.wm.wmexchanger.wmconsumer
+
+import java.io.File
+
+trait RestConsumerish {
+  def open(): Unit
+  def close(): Unit
+  def download(file: File): String
+}
