+++ wmexchanger/src/main/scala/org/clulab/wm/wmexchanger/wmproducer/RestProducerish.scala
@@ -1,0 +1,9 @@
+package org.clulab.wm.wmexchanger.wmproducer
+
+import java.io.File
+
+trait RestProducerish {
+  def open(): Unit
+  def close(): Unit
+  def upload(file: File): String
+}
